Ensure all fdvar domains are unique, start inlining some operations
There's an interesting trade off to be seen here. Right now we seem to arbitrarily copy the domain of an fdvar from one var to another. This way we share memory between variables. That kind of makes sense but can be very confusing when tempted to update a domain inline rather than creating a new instance for it every time. The downside is that we're now going to be copying the domain of every var for every step, even if the var wasn't affected.

Eventually we'll move to a system where we simply have a linked list with deltas. That way for every step of the search, only one domain is cloned and all others have to do a lookup in the chain. We'll see how that goes later.

I've made a start with inlining certain operations but there's many more to be converted similarly.