Refactor reified so that the "tests" are fail-quick read-only calls
There are four (six if you count gt and gte) comparison ops that reified may check. Before it would push the state of the variables and then do a plain propagator step, just to see if it fails. By coding explicitly against that we can often quickly determine whether the propagator would hold. More importantly perhaps, we no longer need to store the state because nothing is changed during the checks.