Eliminate vupid, for now
The vupid (var update id) was a caching mechanism for propagators. But I've eliminated propagators as objects and so they only served to signal whether anything changed. I've slowly changes this to be determined by the called functions and now we're at a point where we don't need the vupid anymore to determine changes. And so we can eliminate it now. Perhaps it will return as a caching mechanism later. For now we aim for simplicity.

Farewell vupid, you've been kind!