Untangle distribution system
This commit attempts to untangle the spaghetti that is the distribution system, solver, and path_solver.

Configuration is now placed on the (root) Space object. As simple as possibles (-> strings). The mirage of function factories is mostly replaced with a simpler lookup system.

Overall this will make the system easier to debug.

Almost full compatibility is accomplished with the old API. Multiverse specs working as intended or fixed where broken (in the reintegration branch).

Various refactorings and small fixes also included that I could not easily separate from this commit.

Tests updated to reflect this change.