Add support for fallback var distributors
When a var distributor has no preference it's now possible to define a fallback distributor to decide on it anyways. This allows you to chain distributors for improved priority governance.

In particular we want a chain from `list > markov > naive` so that we can give an explicit variable priority order for some vars, then prioritize on markov vars in general, and then process the remaining vars in arbitrary order. This makes that possible.