Put the hash lookup for list var dist in config instead of space
With the addition of a fallback config it could happen that the var list distributor occurs multiple times. One "static" list on the root space will not suffice in such case. So instead put the hash on each config or fallback config. That way config clashes should not be possible.

Added a test with "list" as a fallback config for var distribution
Updated all tests to properly pass on the list priority