Use require properly
Gets rid of the `require(path)(FD)` pattern to replace it with regular requires.

Part of the reason was to hopefully circumvent a bug in uglifyjs (but that didnt work), part is to be cleaner, part is to keep circular references in check without demanding specific source orders.

Tests remain to pluck from an FD namespace. This way we can easily switch between testing dev and dist sources, later.

This needs a little more polish but all tests are passing right now.