Mah bad. Missing parenthesis after `to.throw` made those tests a dud
All tests that had `expect(-> foo).to.throw` that wasn't checking for an explicit throw value would not have `()` appended to them, making them effectively worthless. The reason is that a properly lookup by itself is not something that is picked up by the code (exception: getters and setters). So doing `a.foo.throw` is just dead code, it should be `a.foo.throw()`.

This commit fixes all the throw tests, or removes obsolete throw tests, and updates code to make certain other throw tests pass.