Remove root_space and all_var_names from Space in favor of using its config
Essentially the `Space#root_space` property was serving the purpose of `Space#config`. But `config` is a bit more reliable, always present (root or not), and obvious. It will also allow us to separate the tree state from the search node state.

Another bonus is that we no longer juggle the root space in arguments when space was already being passed on.