Fix bug in value list distributor that ignores error value
Internally, the function to find the next first value in the list that occurs in the domain would return a negative constant to indicate that no value was found. This negative value was ignored by the value list distributor and packed in a new domain. Not sure how this doesn't throw up other tests but it didn't.

Added two tests to check this explicitly. Added asserts in the domain creator responsible for this.