Eliminate fdvar as object
This commit wraps up the process of eliminating fdvars as objects. Instead the domain of variables are going to be put inside `space.vardoms` as a simple `name:domain` pair. This simplifies cloning significantly, as well as other processes.

This chain of commits contains a bunch of other changes and a whole set of new tests to support stability.

=============================== Coverage summary ================
Statements   : 97.92% ( 3066/3131 ), 73 ignored
Branches     : 94.65% ( 1415/1495 ), 75 ignored
Functions    : 98.96% ( 285/288 ), 6 ignored
Lines        : 97.97% ( 2508/2560 )
=================================================================