Eliminate numbers as domains/constants in config and most of solver
There were bugs being introduced because of how a number would mean a constant ([value, value]) in one case and a small domain bit-wise flags in other cases. Since it's impossoble to attach meta data to numbers to distinguish them during tests (at least), I've taken a different route and simply banned regular numbers as domain values in all of config and nearly all of the solver. The solver now only accepts numbers (as constants) OR domains for the `decl` method. It also accepts only a number for `num`, where it is hard to be ambiguous.

Updates all the tests because they were explicitly testing this. The tests are still valid because internally the arrays should still be "small" (eg. a number). Just not while in the config.

=============================== Coverage summary ================
Statements   : 96.81% ( 3034/3134 ), 75 ignored
Branches     : 93.26% ( 1412/1514 ), 75 ignored
Functions    : 97.94% ( 285/291 ), 10 ignored
Lines        : 96.96% ( 2454/2531 )
=================================================================