Check for changed domains in propagate() now that all domains are prims
Before domains could be arrays which we couldn't trivially `===`. But now domains are strings and numbers so we can easily do `===` to check whether the domain changed. This is the first step in simplifying some of the propagation code.

Note that this one actually kind of breaks on callback propagators, not really but yeah, however we'll remove that one in the next commit so it's all good.

=============================== Coverage summary ================
Statements   : 91.07% ( 3884/4265 ), 6 ignored
Branches     : 86.66% ( 1644/1897 ), 35 ignored
Functions    : 93.47% ( 358/383 ), 9 ignored
Lines        : 91.28% ( 3307/3623 )
=================================================================