Add DSL version of the einstein problem as a poc example. And improve its setup.
There was one constraint that was missing that reduced the space and runtime. Also the implications (`a->b`) were encoded as `lte` but should be encoded as a biconditional (iif) `a<=>b`, which is an `eq`. This improved runtime dramatically.

=============================== Coverage summary ================
Statements   : 93.2% ( 4564/4897 ), 6 ignored
Branches     : 87.91% ( 1935/2201 ), 40 ignored
Functions    : 95.05% ( 461/485 ), 6 ignored
Lines        : 94.01% ( 3952/4204 )
=================================================================

File dist/finitedomain.dist.min.js created: 214.84 kB → 53.31 kB → 15.94 kB (gzip)