Enable ML_START, implement isall+lte->ltes trick
The ML_START wasn't properly encoded (oops) into the ML

There is a rewrite trick where if you have `A = all?(B C), D <= A` then you can eliminate `A` by rewriting the ops to `D <= B, D <= C`. This works for any number of args in the `isAll`.

This is the first trick that potentially needs more space than the original ops occupy so for this ml-recycling functionality has been added. It will search for jumps that cover at least some space.

When compiling the trick a number of spaces is recycled such that all of them fit the required number of `ML_LTE`'s (the number of args to the `isAll`). Those `ML_LTE`'s are then compiled in, the remainder becomes a new jump. The old ops are then eliminated.

In the case it just concerns 2 vars the existing ops are immediately recycled instead, as is currently the case with most tricks.

=============================== Coverage summary ================
Statements   : 75.63% ( 6171/8160 ), 6 ignored
Branches     : 63.92% ( 2494/3902 ), 23 ignored
Functions    : 77.11% ( 411/533 ), 3 ignored
Lines        : 77.53% ( 5525/7126 )
=================================================================

File dist/finitedomain.dist.min.js created: 353.38 kB → 85.89 kB → 23.65 kB (gzip)